+++ CHANGELOG.md
@@ -7,5 +7,6 @@
 - Add PSN API package
 - Add instructions for authenticaticating with PSN
 - Add method to retrieve PSN API authorisation object using authentication credentials
+- Add method to get an PSN account ID from a username
 
 [unreleased]: https://github.com/yndajas/video-game-library/compare/9b793f6...HEAD